requirements.txt
ansicon==1.89.0
arrow==1.2.3
asgiref==3.6.0
attrs==23.1.0
black==23.3.0
blessed==1.20.0
certifi==2022.12.7
cffi==1.15.1
charset-normalizer==3.1.0
click==8.1.3
colorama==0.4.6
cryptography==40.0.2
defusedxml==0.7.1
Django==4.2
django-allauth==0.54.0
django-picklefield==3.1
django-q==1.3.9
django-rest-framework==0.1.0
djangorestframework==3.14.0
djangorestframework-simplejwt==5.2.2
drf-spectacular==0.26.2
idna==3.4
inflection==0.5.1
jinxed==1.2.0
jsonschema==4.17.3
mypy-extensions==1.0.0
oauthlib==3.2.2
packaging==23.1
pathspec==0.11.1
Pillow==9.5.0
platformdirs==3.2.0
pycparser==2.21
PyJWT==2.6.0
pyrsistent==0.19.3
python-dateutil==2.8.2
python3-openid==3.2.0
pytz==2023.3
PyYAML==6.0
q==2.7
redis==3.5.3
requests==2.28.2
requests-oauthlib==1.3.1
rest-framework-simplejwt==0.0.2
six==1.16.0
sqlparse==0.4.4
tzdata==2023.3
uritemplate==4.1.1
urllib3==1.26.15
django-filter~=23.1
wcwidth==0.2.6
